Move voronoi diagram operation
