+++ src/maps/geo-utils/operators.ts
@@ -14,6 +14,8 @@ import type {
 
 import { BLANK_GEOJSON } from "@/maps/api";
 
+export { geoSpatialVoronoi } from "@/maps/geo-utils/voronoi";
+
 export const safeUnion = (input: FeatureCollection<Polygon | MultiPolygon>) => {
     if (input.features.length === 1) return input.features[0];
     const union = turf.union(input);
